pyproject.toml
[tool.pdm]

[tool.pdm.dev-dependencies]
linting = [
    "black>=23.11.0",
]
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "entsoe-py>=0.5.10",
    "jupyterlab>=4.0.8",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.8.1",
    "jupytext>=1.15.2",
    "ipympl>=0.9.3",
]
requires-python = ">=3.9"
license = {text = "MIT"}
